code is compiling
separated classes into files
