+++ tsconfig.json
@@ -1,13 +1,9 @@
 {
-  "files": [
-    "src/Axiom.ts",
-    "src/Debug.ts",
-    "src/Mathematics.ts",
-    "src/Axiom.d.ts",
-    "src/UIControls/ButtonPagingControl.ts",
-    "src/config/models.ts"
-  ],
   "compilerOptions": {
+    "typeRoots": [
+      "types"
+    ],
+    "rootDir": "src",
     "strict": true,
     "allowJs": true,
     "outDir": "dist/Axiom/",
