+++ types/TE_API_v7.4.d.ts
@@ -735,7 +735,7 @@ interface ISGWorld {
   Version: IVersion;
   Window: IWindow;
   AttachEvent(bstrEventName: string, dispFunc: (...args: any[]) => any): void;
-  DetachEvent(): unknown;
+  DetachEvent(bstrEventName: string, dispFunc: (...args: any[]) => any): void;
   GetOptionParam(): unknown;
   GetParam(...a: any[]): unknown;
   Open(): unknown;
